move requests to basic namespace
