+++ DragonFruit.Common.Data.Tests/Files/FileDownloadTests.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT License. Please refer to the LICENSE file at the root of this project for details
 
 using System.IO;
-using DragonFruit.Common.Data.Utils;
+using DragonFruit.Common.Data.Basic;
 using NUnit.Framework;
 
 namespace DragonFruit.Common.Data.Tests.Files
