+++ DragonFruit.Common.Data/Basic/BasicApiRequestExtensions.cs
@@ -2,16 +2,15 @@
 // Licensed under the MIT License. Please refer to the LICENSE file at the root of this project for details
 
 using System.Collections.Generic;
-using DragonFruit.Common.Data.Utils;
 
-namespace DragonFruit.Common.Data.Extensions
+namespace DragonFruit.Common.Data.Basic
 {
-    public static class BasicRequestExtensions
+    public static class BasicApiRequestExtensions
     {
         /// <summary>
         /// Appends a query parameter to the current <see cref="BasicApiRequest"/>
         /// </summary>
-        public static BasicApiRequest WithQuery(this BasicApiRequest request, string key, object value)
+        public static T WithQuery<T>(this T request, string key, object value) where T : IBasicApiRequest
         {
             return request.WithQuery(key, value.ToString());
         }
@@ -19,7 +18,7 @@ namespace DragonFruit.Common.Data.Extensions
         /// <summary>
         /// Appends a query parameter to the current <see cref="BasicApiRequest"/>
         /// </summary>
-        public static BasicApiRequest WithQuery(this BasicApiRequest request, string key, string value)
+        public static T WithQuery<T>(this T request, string key, string value) where T : IBasicApiRequest
         {
             request.Queries.Value.Add(new KeyValuePair<string, string>(key, value));
             return request;
