+++ DragonFruit.Common.Data/DragonFruit.Common.Data.csproj
@@ -33,8 +33,4 @@
         </None>
     </ItemGroup>
 
-    <ItemGroup>
-      <Folder Include="Requests" />
-    </ItemGroup>
-
 </Project>
